Minor: renamed CreateRoom inot CreateRoomView
